PR 12665: Rename IRequest to IHttpPostMessageRequest and IResponse to IHttpPostMessageResponse
There's no problem when importing the module like that: import * as hpm from 'http-post-message';
However, when you treat hpm as internal module (and that's what we do in PowerBIClients with all libraries), there's no namespace for that module, so we have a global IResponse and IRequest declared in httpPostMessage.d.ts. The problem is that there's already existing declaration of IRequest and IResponse from some other module, so we have an error.